pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "KidBox"
version = "0.1.0"
description = "KidBox launcher and apps"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pygame-ce>=2.5.2",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
